bugFix:修复了dirent和stat的bug
Former-commit-id: cb2170a3947de4de448106d4c67d058e050fe055

+++ fat32_fs/src/vfs.rs
@@ -647,7 +647,7 @@ impl VFile {
 
     // 获取目录中offset处目录项的信息
     // 返回(size, accessed_time, modification_time, creation_time, first_cluster)
-    pub fn stat(&self) -> (u32, u64, i64, i64, i64, u64, u32) {
+    pub fn stat(&self) -> (usize, isize, isize, isize, isize, isize, u32) {
         self.read_short_dirent(|short_ent| {
             let (_, _, _, _, _, _, ctime) = short_ent.get_creation_time();
             let atime = self.accessed_time();
@@ -668,12 +668,12 @@ impl VFile {
                 mode = S_IFLINK;
             }
             (
-                first_cluster,
-                size as u64,
-                atime as i64,
-                mtime as i64,
-                ctime as i64,
-                blocks as u64,
+                first_cluster as usize,
+                size as isize,
+                atime as isize,
+                mtime as isize,
+                ctime as isize,
+                blocks as isize,
                 mode,
             )
         })
